+++ workspaceEmpresa-1.0.0/projeto-empresav1/src/test/java/br/com/contmatic/empresav1/test/TimeoutTestExample.java
@@ -1,0 +1,18 @@
+package br.com.contmatic.empresav1.test;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+public class TimeoutTestExample {
+
+	private boolean Test = true;
+
+	@Test(timeout = 3000)
+	public void teste_timeout_deve_falhar() {
+		// Simulando um simples test de Timeout de 3 segundos
+		assertTrue(Test);
+		while(true);
+			
+	} 
+}
